Update AccessPointsMetrics.php, Controller.php, DevicesMetrics.php, and 4 more files
